法器属性

+++ TaFangSSS/Assets/Scripts/Config/JingJieConfig.cs
@@ -51,16 +51,16 @@ public class JingJieConfig : MonoBehaviour
         { JingJieType.大道圣人, new List<long>() { 10000000000, 20000000000, 30000000000, 50000000000, 100000000000, 200000000000, 300000000000, 500000000000 } },
         { JingJieType.混元圣人, new List<long>() { 30000000000, 50000000000, 80000000000, 120000000000, 200000000000, 300000000000, 500000000000, 1000000000000 } },
     };
-    public static Dictionary<QualityType, decimal> 突破跟脚Dic = new Dictionary<QualityType, decimal>()
+    public static Dictionary<QualityType, float> 突破跟脚Dic = new Dictionary<QualityType, float>()
     {
-        { QualityType.黄品, 1m },
-        { QualityType.玄品, 1.1m },
-        { QualityType.地品, 1.3m },
-        { QualityType.天品, 1.6m },
-        { QualityType.宇品, 2m },
-        { QualityType.宙品, 3m },
-        { QualityType.洪品, 5m },
-        { QualityType.荒品, 10m },
+        { QualityType.黄品, 1f },
+        { QualityType.玄品, 1.1f },
+        { QualityType.地品, 1.3f },
+        { QualityType.天品, 1.6f },
+        { QualityType.宇品, 2f },
+        { QualityType.宙品, 3f },
+        { QualityType.洪品, 5f },
+        { QualityType.荒品, 10f },
     };
     public static Dictionary<JingJieType, int> JingJieAttributeDic =
         new Dictionary<JingJieType, int>()
@@ -86,11 +86,11 @@ public class JingJieConfig : MonoBehaviour
         { JingJieType.鸿蒙,100000},
     };
 
-    public static decimal  跟脚 => Get跟脚();
+    public static float  跟脚 => Get跟脚();
     public static float 轮回系数 = 1;
-    public static decimal  Get跟脚()
+    public static float  Get跟脚()
     {
-        decimal value = PlayerData.S.初始跟脚;
+        float value = PlayerData.S.初始跟脚;
         foreach (var item in PlayerData.S.当前轮回突破Dic)
         {
             if (item.Value != QualityType.None)
